move validate to evaluations module
