+++ aworld/ralph_loop/config.py
@@ -7,11 +7,11 @@ from pydantic import Field
 
 from aworld.config import ModelConfig, TaskConfig, BaseConfig
 from aworld.evaluations.base import EvalCriteria, Scorer
+from aworld.evaluations.reflect import Reflector
 from aworld.ralph_loop.detect.stop_condition import StopCondition
 # from aworld.ralph_loop.mission.analyzer import Analyzer
 # from aworld.ralph_loop.mission.enhancer import ContextEnhancer
 # from aworld.ralph_loop.mission.types import MissionType
-from aworld.ralph_loop.reflect import Reflector
 # from aworld.ralph_loop.plan.base import BasePlanner, BasePlanReviewer, BasePlanOptimizer
 from aworld.ralph_loop.types import ConflictStrategy
